Fix styling of release notes
