+++ UnstoppableWallet/UnstoppableWallet/Modules/Markdown/MarkdownModule.swift
@@ -6,7 +6,7 @@ struct MarkdownModule {
         let provider = MarkdownPlainContentProvider(url: url, networkManager: App.shared.networkManager)
         let service = MarkdownService(provider: provider)
         let parser = MarkdownParser()
-        let viewModel = MarkdownViewModel(service: service, parser: parser)
+        let viewModel = MarkdownViewModel(service: service, parser: parser, parserConfig: AcademyMarkdownConfig.config)
 
         return MarkdownViewController(viewModel: viewModel, handleRelativeUrl: handleRelativeUrl)
     }
@@ -15,7 +15,7 @@ struct MarkdownModule {
         let provider = MarkdownGitReleaseContentProvider(url: url, networkManager: App.shared.networkManager)
         let service = MarkdownService(provider: provider)
         let parser = MarkdownParser()
-        let viewModel = MarkdownViewModel(service: service, parser: parser)
+        let viewModel = MarkdownViewModel(service: service, parser: parser, parserConfig: ReleaseNotesMarkdownConfig.config)
 
         return ReleaseNotesViewController(viewModel: viewModel, handleRelativeUrl: false, urlManager: UrlManager(inApp: false), presented: presented, closeHandler: closeHandler)
     }
