+++ UnstoppableWallet/UnstoppableWallet/Modules/Markdown/MarkdownParser.swift
@@ -4,44 +4,8 @@ import Down
 
 class MarkdownParser {
 
-    private let colors = StaticColorCollection(
-            heading1: .themeOz,
-            heading2: .themeJacob,
-            heading3: .themeJacob,
-            body: .themeBran
-    )
-
-    private let paragraphStyles: StaticParagraphStyleCollection = {
-        var paragraphStyles = StaticParagraphStyleCollection()
-
-        let headingParagraphStyle = NSMutableParagraphStyle()
-
-        let bodyParagraphStyle = NSMutableParagraphStyle()
-        bodyParagraphStyle.lineSpacing = 6
-
-        paragraphStyles.heading1 = headingParagraphStyle
-        paragraphStyles.heading2 = headingParagraphStyle
-        paragraphStyles.heading3 = headingParagraphStyle
-        paragraphStyles.body = bodyParagraphStyle
-
-        return paragraphStyles
-    }()
-
-    func viewItems(content: String, url: URL?, fontSize: Int) -> [MarkdownBlockViewItem] {
-        let fonts = StaticFontCollection(
-                heading1: .title2,
-                heading2: .title3,
-                heading3: .headline2,
-                body: .systemFont(ofSize: CGFloat(fontSize), weight: .regular)
-        )
-
+    func viewItems(content: String, url: URL?, configuration: DownStylerConfiguration) -> [MarkdownBlockViewItem] {
         let down = Down(markdownString: content)
-
-        let configuration = DownStylerConfiguration(
-                fonts: fonts,
-                colors: colors,
-                paragraphStyles: paragraphStyles
-        )
 
         let styler = DownStyler(configuration: configuration)
